projects/argilla-sdk-bot/requirements_docs_dataset.txt
#markdown-it-py==3.0.0
PyGithub==2.3.0
#nltk==3.8.1
datasets==2.19.2
unstructured==0.14.5